+++ src/pages/_app.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import type { AppProps } from 'next/app';
 import { ThemeProvider } from 'styled-components';
 import defaultTheme from '@/styles/theme';
