+++ src/pages/index.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Header } from '../components/Header';
 
 const Home = () => {
